Add ClientIpTrait for standardized client IP extraction in authentication handlers
