+++ src/Handlers/Auth/LogoutHandler.php
@@ -35,6 +35,8 @@ use Psr\Http\Message\ServerRequestInterface;
  */
 final class LogoutHandler extends AbstractHandler
 {
+    use ClientIpTrait;
+
     private ?JwtService $jwtService = null;
     private Logger $authLogger;
 
@@ -72,39 +74,6 @@ final class LogoutHandler extends AbstractHandler
             $this->jwtService = JwtServiceFactory::fromEnv();
         }
         return $this->jwtService;
-    }
-
-    /**
-     * Get the client IP address, checking proxy headers first.
-     *
-     * Checks X-Forwarded-For and X-Real-IP headers (common in reverse proxy setups)
-     * before falling back to REMOTE_ADDR. For X-Forwarded-For, uses the first IP
-     * in the chain (the original client).
-     *
-     * @param ServerRequestInterface $request The incoming HTTP request.
-     * @param array<string, mixed> $serverParams Server parameters from the request.
-     * @return string The client IP address, or 'unknown' if not determinable.
-     */
-    private function getClientIp(ServerRequestInterface $request, array $serverParams): string
-    {
-        // Check X-Forwarded-For header (may contain comma-separated list of IPs)
-        $forwardedFor = $request->getHeaderLine('X-Forwarded-For');
-        if ($forwardedFor !== '') {
-            // Use the first IP in the chain (original client)
-            $ips = array_map('trim', explode(',', $forwardedFor));
-            if (!empty($ips[0])) {
-                return $ips[0];
-            }
-        }
-
-        // Check X-Real-IP header
-        $realIp = $request->getHeaderLine('X-Real-IP');
-        if ($realIp !== '') {
-            return $realIp;
-        }
-
-        // Fall back to REMOTE_ADDR
-        return is_string($serverParams['REMOTE_ADDR'] ?? null) ? $serverParams['REMOTE_ADDR'] : 'unknown';
     }
 
     /**
@@ -164,8 +133,8 @@ final class LogoutHandler extends AbstractHandler
                 $jwtService = $this->getJwtService();
                 $username   = $jwtService->extractUsername($token) ?? 'unknown';
             } catch (\Throwable $e) {
-                // Any JWT/config issue: keep logout successful, log at debug level
-                $this->authLogger->debug('Failed to extract username from JWT during logout', [
+                // Any JWT/config issue: keep logout successful, log for visibility
+                $this->authLogger->info('Failed to extract username from JWT during logout', [
                     'error' => $e->getMessage()
                 ]);
             }
